HAL: Change idle loop instruction from NOP to WFI for ES3
Change the idle loop instruction from NOP(no operation) to
WFI(wait for interrupt) if built for ES3. ES3 fixed the
crashing issue when the idle loop use WFI in the debug
build and GDB is attached to board. Replacing NOP with WFI
in the idle can reduce the power consumption when the CPU
is idle.

Testing Done: Developer tested on both DB3/ES2 and DB3/ES3
boards with debug build and GDB attached to board.
Bug Number: SW-2154
Review URL: TBD
Reviewed by: TBD

Signed-off-by:  Kim Mui<kimm@bsquare.com>